test we can improve tax configuration
